+++ interface/include/nakama-cpp/realtime/wslay/WslayIOInterface.h
@@ -14,8 +14,8 @@ enum class NetIOAsyncResult {
 class WslayIOInterface {
 public:
     virtual ~WslayIOInterface() {}
-    virtual ssize_t recv(void* buf, size_t len, int* wouldBlock) = 0;
-    virtual ssize_t send(const void* data, size_t len, int* wouldBlock) = 0;
+    virtual int recv(void* buf, size_t len, int* wouldBlock) = 0;
+    virtual int send(const void* data, size_t len, int* wouldBlock) = 0;
     virtual void close() = 0;
     virtual NetIOAsyncResult connect_init(const URLParts& urlParts) = 0;
     virtual NetIOAsyncResult connect_tick() = 0;
